# app/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "description": "sean",
    "version": "1.0.0",
    "title": "Seans-Python3-Flask-Rest-Boilerplate",
    "contact": {
      "email": "seanwasere@gmail.com"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/"
    }
  ],
  "tags": [
    {
      "name": "product Request",
      "description": "Example API for requesting and return product requests"
    }
  ],
  "paths": {
    "/api/product": {
      "get": {
        "tags": [
          "product Request"
        ],
        "summary": "Returns productRequests",
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/productRequest"
            }
          }
        }
      },
      "post": {
        "tags": [
          "product Request"
        ],
        "summary": "Create a new product request system",
        "requestBody": {
          "description": "product Request Post Object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/productRequestPostBody"
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "201": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/id"
            }
          },
          "400": {
            "description": "Failed. Bad post data."
          }
        }
      }
    },
    "/api/product/{id}": {
      "parameters": [
        {
          "name": "id",
          "in": "path",
          "required": true,
          "description": "ID of the cat that we want to match",
          "type": "int"
        }
      ],
      "get": {
        "tags": [
          "product Request"
        ],
        "summary": "product book request with given ID",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "product Request id",
            "schema": {
              "$ref": "#/components/schemas/id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/bookRequest"
            }
          },
          "400": {
            "description": "Failed. Misunderstood Request."
          },
          "404": {
            "description": "Failed. product request not found."
          }
        }
      },
      "put": {
        "summary": "edit a product request by ID",
        "tags": [
          "product Request"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "product Request id",
            "schema": {
              "$ref": "#/components/schemas/id"
            }
          }
        ],
        "requestBody": {
          "description": "product Request Object",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/productRequestPostBodyEdit"
              }
            }
          }
        },
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/productRequest"
            }
          },
          "400": {
            "description": "Failed. Bad post data."
          }
        }
      },
      "delete": {
        "summary": "Delete product Request by ID",
        "tags": [
          "product Request"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "description": "product Request Id",
            "schema": {
              "$ref": "#/components/schemas/id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "OK",
            "schema": {
              "$ref": "#/components/schemas/id"
            }
          },
          "400": {
            "description": "Failed. Misunderstood Request."
          },
          "404": {
            "description": "Failed. product Request not found."
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "id": {
        "properties": {
          "id": {
            "type": "int"
          }
        }
      },
      "productRequestPostBody": {
        "type": "object",
        "properties": {
          "codeOfProduct": {
            "type": "string",
            "format": "string"
          },
          "name": {
            "type": "string",
            "format": "string"
          },
          "typeOfProduct": {
            "type": "string",
            "format": "string"
          },
          "is_product": {
            "type": "string",
            "format": "string"
          },
          "count": {
            "type": "string",
            "format": "string"
          },
          "price": {
            "type": "string",
            "format": "string"
          },
          "description": {
            "type": "string",
            "format": "string"
          }
        }
      },
       "productRequestPostBodyEdit": {
        "type": "object",
        "properties": {
          "codeOfProduct": {
            "type": "string",
            "format": "string"
          },
          "name": {
            "type": "string",
            "format": "string"
          },
          "typeOfProduct": {
            "type": "string",
            "format": "string"
          },
          "is_product": {
            "type": "string",
            "format": "string"
          },
          "count": {
            "type": "string",
            "format": "string"
          },
          "price": {
            "type": "string",
            "format": "string"
          },
          "description": {
            "type": "string",
            "format": "string"
          }
        }
      },
      "productRequest": {
        "type": "object",
        "properties": {
          "codeOfProduct": {
            "type": "string",
            "format": "string"
          },
          "name": {
            "type": "string",
            "format": "string"
          },
          "typeOfProduct": {
            "type": "string",
            "format": "string"
          },
          "is_product": {
            "type": "string",
            "format": "string"
          },
          "count": {
            "type": "string",
            "format": "string"
          },
          "price": {
            "type": "string",
            "format": "string"
          },
          "description": {
            "type": "string",
            "format": "string"
          }
        }
      }
    }
  }
}